+++ CHANGELOG.md
@@ -13,6 +13,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+## [1.127.0]
+
+### Added
+- Constructor-schemes: added dynamic-image node
+- onDataCompare added ability to compare cells with cells
+- Visualizations. Added the ability to download the result in xlsx
+
+### Fixed
+- Storage data quota exceeded
+- Constructor-schemes: incorrect save scheme to store
+
 ## [1.126.0]
 
 ### Added